Split boombox pipeline into bin and pipeline
